% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aggregate_SWAT_ASC.R
\name{Aggregate.SWAT.ASC}
\alias{Aggregate.SWAT.ASC}
\title{Non exportet aggregate function over profile}
\usage{
Aggregate.SWAT.ASC(SWATi, soil)
}
\arguments{
\item{SWATi}{swati from swatday.asc}

\item{soil}{soil.df}
}
\value{
Returns the desired output to the database directly.
}
\description{
...
}
