% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EvapDailyToDailyAgg.R
\name{Evap.DailyToDailyAgg}
\alias{Evap.DailyToDailyAgg}
\title{Internal Helper Evap.DailyToDailyAgg}
\usage{
Evap.DailyToDailyAgg(dat)
}
\arguments{
\item{...}{}
}
\value{
Returns the desired output
}
\description{
...
}
