% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EvapDailyToMonthly.R
\name{Evap.DailyToMonthly}
\alias{Evap.DailyToMonthly}
\title{Internal Helper Evap.DailyToMonthly}
\usage{
Evap.DailyToMonthly(dat)
}
\arguments{
\item{...}{}
}
\value{
Returns the desired output
}
\description{
...
}
