% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EvapDailyToVegper.R
\name{Evap.DailyToVegper}
\alias{Evap.DailyToVegper}
\title{Internal Helper Evap.DailyToVegper}
\usage{
Evap.DailyToVegper(dat, vp.year, vp.start, vp.end)
}
\arguments{
\item{...}{}
}
\value{
Returns the desired output
}
\description{
...
}
