% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EvapDailyToYearly.R
\name{Evap.DailyToYearly}
\alias{Evap.DailyToYearly}
\title{Internal Helper Evap.DailyToYearly
...}
\usage{
Evap.DailyToYearly(dat)
}
\arguments{
\item{...}{}
}
\value{
Returns the desired output
}
\description{
Internal Helper Evap.DailyToYearly
...
}
