% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlowDailyToDailyAgg.R
\name{Flow.DailyToDailyAgg}
\alias{Flow.DailyToDailyAgg}
\title{Internal Helper Flow.DailyToDailyAgg}
\usage{
Flow.DailyToDailyAgg(dat, bypar)
}
\arguments{
\item{...}{}
}
\value{
Returns the desired output
}
\description{
...
}
