% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlowDailyToMonthly.R
\name{Flow.DailyToMonthly}
\alias{Flow.DailyToMonthly}
\title{Internal Helper Flow.DailyToMonthly}
\usage{
Flow.DailyToMonthly(dat)
}
\arguments{
\item{...}{}
}
\value{
Returns the desired output
}
\description{
...
}
