% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlowDailyToYearly.R
\name{Flow.DailyToYearly}
\alias{Flow.DailyToYearly}
\title{Internal Helper Flow.DailyToYearly}
\usage{
Flow.DailyToYearly(dat)
}
\arguments{
\item{...}{}
}
\value{
Returns the desired output
}
\description{
...
}
