% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlowMonthlyToYearly.R
\name{Flow.MonthlyToYearly}
\alias{Flow.MonthlyToYearly}
\title{Internal Helper Flow.MonthlyToYearly}
\usage{
Flow.MonthlyToYearly(dat, bypar)
}
\arguments{
\item{...}{}
}
\value{
Returns the desired output
}
\description{
...
}
