% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnc_add_nFK.R
\name{fnc_add_nFK}
\alias{fnc_add_nFK}
\title{Calculation of the plant available water content in \% and mm}
\usage{
fnc_add_nFK(df)
}
\arguments{
\item{df}{is the list of soil properties created by \code{\link{fnc_get_soil}}}
}
\value{
returns the same list of soil properties with columns containing AWC-infos added
}
\description{
This function quickly adds nFK in \% and mm to a list of soil data frames
}
