% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnc_app.R
\name{fnc_app}
\alias{fnc_app}
\title{Central Application}
\usage{
fnc_app(
  df.ids,
  how_many_points_at_once = 1000,
  start_date,
  end_date,
  treespecies,
  wanted_daily = NA,
  wanted_layer = NA,
  db_name,
  meta = NA,
  ...
)
}
\arguments{
\item{df.ids}{a data frame containing the name of the ids \code{ID}, as well as the UTM-32632-Coordinates in columns named \code{easting} and  \code{northing}.}

\item{how_many_points_at_once}{depending on the user's RAM, the function runs 1000 points at once by default (proved useful on a 8GB RAM machine). Can be changed manually if more RAM is available.}

\item{start_date}{start date as \code{Date}-object}

\item{end_date}{end date as \code{Date}-object}

\item{treespecies}{name of the tree species to be modelled with. Either a single species name that is then used for all points in \code{df.ids}, or a vector of the same length as \code{df.ids$ID} if the main tree species of each point is known and differs. Must be one of \code{beech}, \code{oak}, \code{spruce}, \code{pine}, \code{larch}, \code{douglasfir}.}

\item{db_name}{name and file path of the SQL-database}

\item{meta}{\code{NA} by default, a location where meta-information can be written. Will then be written into the database}

\item{...}{all additional information that can be set in \code{\link{fnc_get_soil}}. See help page for all options that can be selected}

\item{layercols}{a sting containing the desired output products. Full list of possible output columns can be find on the help page of \code{\link[LWFBrook90R]{run_LWFB90}} under \code{Layer outputs}}

\item{dailycols}{same as layercols but for daily output. For possible options see \code{\link[LWFBrook90R]{run_LWFB90}}}
}
\value{
An SQLite Database with the tables soil, and daily and/or layer outputs.
}
\description{
Work in progress-version of a function that makes the application extremely easy for final users.
}
