% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnc_AWC.R
\name{fnc_AWC}
\alias{fnc_AWC}
\title{Calculation of the plant available water content}
\usage{
fnc_AWC(data, nFK = F)
}
\arguments{
\item{data}{is the list of soil properties calculated by \code{\link{fnc_get_soil}}}

\item{nFK}{indicates whether the AWC is calculated for 1 m soil depth (\code{nFK = T}), returning a vector with one vector element for each list element. \code{nFK = F} returns the initial list with the additional columns: \code{fc} (vol. water content at 63 hPa), \code{pwp} (vol. water content at 15000 hPa) and \code{AWC} (AWC in mm per m^2 and depth of the soil layer)}
}
\value{
Returns either the expanded input list of soil properties or a vector containing the AWC for the whole soil depth with one vector element per input list element
}
\description{
This function calculates the plant available water content (AWC, nutzbare Feldkapazitaet) from the Mualem-van Genuchten retention curve unsing \code{\link{fnc_MvG.swc}}. The function calculates the AWC per soil layer or for the whole depth at once. The gravel content is already taken into account.
}
