% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnc_create_IDs.R
\name{fnc_create_IDs}
\alias{fnc_create_IDs}
\title{Create ID-raster pattern within polygon with set resolution}
\usage{
fnc_create_IDs(
  poly,
  res = 250,
  ID_pre = "ID_",
  out_dir,
  out_name = "IDs",
  add_tranches = F
)
}
\arguments{
\item{poly}{polygon of the test area in EPSG:32632 as full path with file name and ending. Accepts all files that can be read with \code{sf::st_read}}

\item{res}{resolution in m}

\item{ID_pre}{optional, prefix for ID_custom-name as character. Default is \code{ID_}}

\item{out_dir}{mandatory, directory where df.ids shall be stored.}

\item{out_name}{optional, name for ID-files, default is \code{IDs.rds}}

\item{add_tranches}{when large areas are modelled with a high resolution, it can be useful to model your area as tranches. In this case the IDs are assigned to 9 Tranches, similar to a SUDOKU-field. This way you can calculate the results tranche by tranche and in a 9th of the total computing time you get a result covering the whole modelling area. If \code{TRUE}, df.ids are stored as 9 tranches named \code{out_name _ trX.rds} in \code{out_dir}.}
}
\value{
returns a list of points with ID_custom (optional), x- and y- coordinates
}
\description{
This function creates a regular pattern of modelling points for a given test area and a given resolution.
}
