% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnc_depth_disc.R
\name{fnc_depth_disc}
\alias{fnc_depth_disc}
\title{Create depth discretisation for soil data frame}
\usage{
fnc_depth_disc(
  df,
  limit_bodtief = NA,
  thick_1 = 5,
  thick_2 = 10,
  thick_3 = 20,
  disk_gr1 = 50,
  disk_gr2 = 100
)
}
\arguments{
\item{df}{data frame containing mandatory depth information on horizon number (\code{mat}), \code{upper} and \code{lower} as columns. Information of soil physics are optional and will be kept and returned.}

\item{limit_bodtief}{max soil depth, default is \code{NA} and uses max soil depth as defined in \code{df.LEIT}. If not \code{NA} soil-dfs are created down to the depth specified here as depth in \code{m}, negative. Might be used to give room for different \code{maxrootdepth} - settings in \link{fnc_get_params}. In this case, soil depth may be reduced significantly.}

\item{thick_1}{first set distance between nodes the function will build down to \code{disk_gr1}}

\item{thick_2}{second set distance between nodes the function will build down to \code{disk_gr2}}

\item{thick_3}{third set distance between nodes the function will build below \code{disk_gr2}}

\item{disk_gr1}{border of the first thickness change form \code{thick_1} to \code{thick_2}}

\item{disk_gr2}{border of the second thickness change form \code{thick_2} to \code{thick_3}}
}
\value{
A soil - data frame with a set position of soil nodes.
}
\description{
This function takes a soil data frame that contains the depth and thickness of the soil horizons and
creates a soil data frame that creates n modelling nodes with \code{thick_1} distance from the top to \code{disk_gr1},
n2 nodes with \code{thick_2} distance from the top to \code{disk_gr2}, and n3 nodes with \code{thick_3} distance below \code{disk_gr2},
while at the same time keeping the position of the soil horizons. Distances between nodes and soil horizon depths
that do not add up to another layer are added to the layer above.
}
\examples{
df.soil <- data.frame(
  "mat" = 1:4,
  "upper" = c(0,13,60,101),
  "lower" = c(13,60,101,135),
  "texture" = c("Ls2", "Ls3", "Ls3", "Lts")
)

df <- df.soil
fnc_depth_disc(df = df.soil)

}
