% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnc_extract_points_from_rstack_bze.R
\name{fnc_extract_points_bze}
\alias{fnc_extract_points_bze}
\title{Extract BZE information from raster stack}
\usage{
fnc_extract_points_bze(lay, xy, meta.out, buffering = F, buff_width = 50)
}
\arguments{
\item{lay}{rasterstack}

\item{xy}{Spaitalpointsdataframe containing points in UTM25832}

\item{buffering}{should most common value in \code{buff_width} buffer distance be used if \code{NA}}

\item{buff_width}{buffer width}
}
\value{
Set of soil information that is further processed in function fnc_soil_bze()
}
\description{
A function written by Paul Schmidt-Walter that extracts soil data from a rasterstack of regionalised BZE-Data. If data is missing, it takes most common value within a \code{buff_width} buffer.
}
