% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnc_extract_points_from_rstack_dgm.R
\name{fnc_extract_points_dgm}
\alias{fnc_extract_points_dgm}
\title{Extract slope and aspect from dgm}
\usage{
fnc_extract_points_dgm(lay, xy, buffering = F, buff_width = 50)
}
\arguments{
\item{lay}{rasterstack}

\item{xy}{Spaitalpointsdataframe containing points in GK3}

\item{buffering}{should most common value in \code{buff_width} buffer distance be used if \code{NA}}

\item{buff_width}{buffer width}
}
\value{
Set of soil information that is further processed in function fnc_soil_bze()
}
\description{
Extract slope and aspect from dgm
}
