% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnc_fill_bze_reg.R
\name{fnc_fill_bze_reg}
\alias{fnc_fill_bze_reg}
\title{Filling and reducing the BZE_reg-rasters}
\usage{
fnc_fill_bze_reg(
  which_rasters = NA,
  cells_belt = 6,
  distance_thres = 400,
  verbose = T
)
}
\arguments{
\item{cells_belt}{numeric. Number of cells of moving window. The bigger, the coarser the mean value of the filled cells.}

\item{distance_thres}{numeric. Maximum distance that should be covered by gap-filling-process.}

\item{verbose}{shall current raster in progress be printed to the console}

\item{which_raster}{selection of rasters to be processed. Must be character string with ".tif" endings. With default \code{NA}, all rasters in \code{path_models_in} will be processed.}
}
\value{
updates the filled BZE_reg rasters
}
\description{
This function takes the BZE-raster and fills potentially missing data to the current extent of forest in BW. Cells are filled with a focal moving window approach. Furthermore, small bits of unconnected forest below 1 ha are deleted. Those are usually parks and hedges. Filled rasters are stored under the same name in a different folder that can be changed if necessary.
}
