% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnc_filter.R
\name{fnc_filter}
\alias{fnc_filter}
\title{Short filter function to filter the in-between-ness of df-rows regarding vectors}
\usage{
fnc_filter(df, depths)
}
\arguments{
\item{df}{data frame with \code{upper} and \code{lower}}

\item{depths}{target depth}
}
\value{
filtered dataframe
}
\description{
Filtering multiple rows according to multiple target depths
}
