% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnc_get_params.R
\name{fnc_get_params}
\alias{fnc_get_params}
\title{Parameter-list creation}
\usage{
fnc_get_params(df.ids, tree_species, df.ind.info = NULL)
}
\arguments{
\item{df.ids}{a data frame containing the following columns:
\itemize{
\item \code{ID_custom} - a unique ID-column for assignment that all intermediate products as well as the output will be assigned to.
\item \code{easting} and \code{northing} - coordinates in UTM EPSG:32632
}}

\item{tree_species}{name of the tree species to be modelled with. Either a single species name that is then used for all points in \code{df.ids}, or a vector of the same length as \code{df.ids$ID} if the main tree species of each point is known and differs. Must be one of \code{beech}, \code{oak}, \code{spruce}, \code{pine}, \code{larch}, \code{douglasfir}. \cr It is  recommended (findings of project WHH-KW) to use the settings of spruce for fir and the settings of pine for larch. If certain tree specific parameters should be changed permanently, the corresponding object \code{params_...} should be overwritten manually before running the function.}

\item{df.ind.info}{a data frame containing individual site information for each ID or if parameters are presumed to be different than the default settings of \code{\link[LWFBrook90R]{set_paramLWFB90}}. They need to be given here in the form of a data frame containing the column \code{ID}, which should be identical to the \code{ID}-column of\code{df.ids}, and additional columns that are named exactly like the parameters in \code{\link[LWFBrook90R]{set_paramLWFB90}}.}
}
\value{
Returns a list of parameter settings that can be read and further processed by \code{\link[LWFBrook90R]{run_multisite_LWFB90}} or \code{\link[LWFBrook90R]{run_LWFB90}}
}
\description{
This function creates a list of parameter settings. Vegetation settings, by default, represent the findings of the WHH-KW project.
}
\examples{
ls.param <- fnc_get_params(df.ids = test.ids.bds,
                            tree_species = c("beech", "spruce", "pine", "douglasfir", "beech"))
# as we're working with a list of long lists. For comparative reasons, this function returns the n-th object of each list element
fun1 <- function(lst, n){
  sapply(lst, `[`, n)
}

# tree-species
fun1(ls.param, which(names(ls.param[[1]]) == "budburst.species"))
fun1(ls.param, which(names(ls.param[[1]]) == "glmax"))
fun1(ls.param, which(names(ls.param[[1]]) == "coords_x"))

# all oaks, but change height and age of stands at begin of simulation (parameter "height" & "age.ini")
df.infos <- data.frame("ID_custom" = LETTERS[5:1],
                  "height" = c(10, 15, 30, 20, 20),
                  "age.ini" = c(10, 20, 25, 70, 90))

ls.param <- fnc_set_params(df.ids = test.ids.bds,
                           tree_species = "oak",
                           df.ind.info = df.infos)
# tree-species
fun1(ls.param, which(names(ls.param[[1]]) == "budburst.species"))
fun1(ls.param, which(names(ls.param[[1]]) == "height"))
fun1(ls.param, which(names(ls.param[[1]]) == "age.ini"))

}
