% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnc_get_stok_data.R
\name{fnc_get_stok_data}
\alias{fnc_get_stok_data}
\title{Helper function for selecting STOK_IDs for pre-processed STOK soil data.}
\usage{
fnc_get_stok_data(poly, PTF)
}
\arguments{
\item{poly}{polygon of the test area in EPSG:32632 as full path with file name and ending. Accepts all files that can be read with \code{sf::st_read}}
}
\value{
returns a list of data frames with ID_custom (optional), x- and y- coordinates
}
\description{
Other than with regionalised soil data, there is a limited amount of polygons in the Standortskartierung (~500.000). To speed things up for individual applications, all those have been calculated with all three available PTFs. This function checks, which STOK-polyogns are within the boundaries of the test area, selects the respective soil-dfs, and returns a list.
}
