% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnc_limit_MvG.R
\name{fnc_limit}
\alias{fnc_limit}
\title{Restriction of Mualem-vanGenuchten Values}
\usage{
fnc_limit(df)
}
\arguments{
\item{df}{a data frame containing soil properties including MvG-alpha (column \code{alpha}), MvG-alpha (column \code{alpha}), stone vol_\% (column \code{gravel}), MvG-npar (column \code{npar}), saturated conductivity (column \code{ksat}), and tortuousity (column \code{tort}).}
}
\value{
returns the same data frame, just with parameters limited as described in the function.
}
\description{
The function fnc_PTF creates Mualem-vanGenuchten parameters from physical soil properties. Since some of these PTFs may extrapolate to values that are unrealistic, this function, if activated through \code{limit_mvG} in \code{fnc_create_boden}(), limits the parameters to parameter ranges as described in "U:\\Brook90_2018\\paul_schmidt_walter_2018\\Dokumentation\\2_Bodenparameter.nb".
}
\details{
ranges are as follows:
\itemize{
\item alpha - between 0.1 and 500 m-1
\item gravel - max 95\% if gravel > 95\%
\item npar - max 2 if npar > 2
\item ksat - min 0.01 mm/d if ksat < 0.01
\item tort - between -3 and 6
}
}
