% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnc_make_ssc.R
\name{fnc_make_ssc}
\alias{fnc_make_ssc}
\title{Sand-Silt-Clay from texture class}
\usage{
fnc_make_ssc(texture)
}
\arguments{
\item{texture}{must be one of \code{as.data.frame(soiltexture::TT.classes.tbl("DE.BK94.TT"))[,1]}}
}
\value{
Percentage value of Sand-Silt-Clay in that order.
}
\description{
This simple  creates a grain size distribution as a percentage vector from the texture class as found in Bodenkundliche Kartieranleitung (KA5), p142. The distribution is derived from the centre of the respective polygon in the soil texture triangle.
}
\examples{
make_ssc("Lt3")
}
\references{
Ad-Hoc-Arbeitsgruppe Boden (2005). Bodenkundliche Kartieranleitung. 5. Auflage, Hannover.
}
