% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnc_relateCoords.R
\name{fnc_relateCoords}
\alias{fnc_relateCoords}
\title{Relate coordinates to locations in climatic database}
\usage{
fnc_relateCoords(df.ids, path_std = "R:/klima/whh/brook90_input/locations")
}
\arguments{
\item{df.ids}{a data frame containing the following columns:
\itemize{
\item \code{ID} - a numbered ID that is created in fnc_get_clim for unique assignment within functions
\item \code{ID_custom} - a unique ID-column for assignment that all intermediate products as well as the output will be assigned to.
\item \code{easting} and \code{northing} - coordinates in UTM EPSG:32632
}}

\item{path_std}{path to standard locations directory}
}
\value{
A data frame containing:

\itemize{
\item id and coordinates of custom location and nearest standard location in CRS=EPSG:32632.
}

\itemize{
\item membership of location in tranche.
}
}
\description{
Relate custom locations of interest to standard locations present in a climatic database by nearest neighbour operation.

The standard locations represent the center points of underlying climatic raster data cells located in forest areas of Baden-Württemberg.
}
\details{
Spatial resolution of the underlying raster data and therefore the grid of standard locations is 250 m. Coordinates reference system of the raster data is WGS84-UTM32N, EPSG:32632.

The climatic database provides daily information for several climatic variables at standard locations as input for the Brook90-model. The population of almost 230.000 standard locations is spatialised to 9 tranches, each tranche including approximately 25.000 locations.

Raster representation in the relation of custom to standard location is ensured \code{Value = 1}, as long as the custom location is inside the raster cell underlying the standard location. Therefore the distance from custom location to nearest standard location has to be <= sqrt((250/2)^2*2)) \link{m}.
}
\examples{

fnc_relateCoords(df.ids = test.ids.bds)
}
\references{
Dietrich, H., Wolf, T., Kawohl, T., Wehberg, J., Kändler, G., Mette, T., Röder, A., Böhner, J. (2019): “Temporal and spatial high-resolution climate data from 1961 to 2100 for the German National Forest Inventory (NFI)”, Annals of Forest Sience, 76:6. https://doi.org/10.1007/s13595-018-0788-5
}
\author{
Thilo Wolf \href{mailto:thilo.wolf@forst.bwl.de}{thilo.wolf@forst.bwl.de}, adjusted by Raphael Habel \href{mailto:raphael.habel@forst.bwl.de}{raphael.habel@forst.bwl.de}
}
