% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnc_soil_bze.R
\name{fnc_soil_bze}
\alias{fnc_soil_bze}
\title{Soil-list creation from BZE data}
\usage{
fnc_soil_bze(df.gk)
}
\arguments{
\item{df.gk}{A spatialpointsdataframe with the desired points in GK-3.}
}
\value{
Returns a list of soil data frames.
}
\description{
This function is a wrapper of several smaller functions and chunks of code, all retrieved from "U:\\Brook90_2018\\paul_schmidt_walter_2018\\Dokumentation\\2_Bodenparameter.nb". Combining all this code into one function, it takes a spatialpointsdataframe of coordinates in GK-3 and returns a list of soil data frames. Those are further processed in \code{\link{fnc_create_soil}} by adding soil hydraulic information, humus, and fine roots and can then be read by \code{\link[LWFBrook90]{msiterunLWFB90}}.
}
