% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnc_soil_stok.R
\name{fnc_soil_stok}
\alias{fnc_soil_stok}
\title{Soil-list creation from STOKA data}
\usage{
fnc_soil_stok(df, df.LEIT)
}
\arguments{
\item{df}{A data frame containing the columns \code{ID} as distinct assignment from the initial \code{df.ids} dataframe and the column \code{RST_F}, which is created in \code{\link{fnc_create_soil}} through a spatial join of the coordinates with the STOK-BW shapefile (or, at this stage, the shapefile of test areas).}

\item{df.LEIT}{a data frame containing LEITPROFILE. At this stage of development, the comprehensive BW-wide database is not complete yet, so the test area needs to be assigned here. However, this is done automatically in  \code{\link{fnc_create_soil}}.}
}
\value{
Returns a list of soil data frames.
}
\description{
This function transforms data as currently stored in the LEITPROFIL-database and returns a list of soil data frames as recognised by \code{LWFBrook90R}. The list is further processed in \code{\link{fnc_create_soil}} by adding soil hydraulic information, humus, and fine roots and can then be read by \code{\link[LWFBrook90]{msiterun}}.
}
