% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnc_transf_to_gk.R
\name{fnc_transf_to_gk}
\alias{fnc_transf_to_gk}
\title{UTM-GK3 transformation function}
\usage{
fnc_transf_to_gk(df)
}
\arguments{
\item{df}{a data frame containing the name of the ids \code{ID}, as well as the UTM-Coordinates in columns named \code{easting} and  \code{northing}.}
}
\value{
A spatialpointsdataframe with the projected points in GK3. Data type changes because of further processing within fnc_soil_bze( ).
}
\description{
Some of the data (i.e. DGM, regionalised BZE-data) is stored at the FVA in rasters with GK-3 projection. This function takes a dataframe of Coordinates in UTM (EPSG:32632) and projects them to GK-3 for data extraction of GK-3 rasters.
}
\examples{
test.ids.utm <- data.frame("ID" = c(1,2, 3, 4, 5),
                           "easting" = c(493497, 503000, 515138, 501000, 503330),
                           "northing" = c(5281811, 5292500, 5289355, 5293600, 5291700))

fnc_transf_to_gk(df = test.ids.utm)
}
