% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnc_write_agg.R
\name{fnc_write_agg}
\alias{fnc_write_agg}
\title{Function to aggregate and write data from automated LWFB90-Runs}
\usage{
fnc_write_agg(x, aggr_tp, col_select_vp = NA, col_select_mon = NA, db_name)
}
\arguments{
\item{x}{one of the intermediate producs of \code{\link[LWFBrook90R]{run_LWFB90}} or  \code{\link[LWFBrook90R]{run_multisite_LWFB90}}, which is further processed internally. Can't be adjusted.}

\item{aggr_tp}{a string containing the desired aggregation time period. Can be \code{monthly},  \code{vegper}, or  \code{monthly_vegper}. The latter does both.}

\item{col_select_vp}{a string containing the desired columns from the vegperiod-aggregation (see details)}

\item{col_select_mon}{a string containing the desired columns from the monthly-aggregation (see details)}

\item{db_name}{name and file path of the SQL-database}
}
\value{
Returns the desired output to the database directly.
}
\description{
LWFBrook90 creates a lot of output files. In order to keep data storage to a minimum, both \code{\link[LWFBrook90R]{run_LWFB90}} and \code{\link[LWFBrook90R]{run_multisite_LWFB90}} provide an \code{output_fun} - argument that can be used to reduce the output and directly write it to a database. This is what this function is made for. \cr In comparison to \code{\link{fnc_write}}, which only reduces the columns returned by \code{\link[LWFBrook90R]{run_LWFB90}} (see help page), this function enables aggregation over vegperiod and monthly, plus a more detailed selection of drought indices. See detail section.
}
\section{Vegperiod outputs}{

\tabular{llcl}{
\strong{Name} \tab \strong{Description} \tab \strong{Unit} \cr
tran \tab transpiration \tab mm \cr
irvp \tab evaporation of intercepted rain \tab mm \cr
isvp \tab evaporation of intercepted snow \tab mm \cr
slvp \tab year \tab - \cr
ptran \tab potential transpiration \tab mm \cr
pslvp \tab potential soil evaporation \tab mm \cr
...
}
}

