% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnc_write_climdb.R
\name{fnc_write_climdb}
\alias{fnc_write_climdb}
\title{Creation of Climate Database}
\usage{
fnc_write_climdb(
  df.ids,
  mindate = as.Date("2010-01-01"),
  maxdate = as.Date("2011-12-31"),
  path_std = "R:/klima/whh/brook90_input/locations",
  path_climdb = "R:/klima/whh/brook90_input/db/",
  dbout_name,
  append_to_db = F
)
}
\arguments{
\item{df.ids}{a data frame containing the following columns:
\itemize{
\item \code{ID_custom} - a unique ID-column for assignment that all intermediate products as well as the output will be assigned to.
\item \code{easting} and \code{northing} - coordinates in UTM EPSG:32632
}}

\item{mindate}{first day of modelling time period as \code{Date}- object}

\item{maxdate}{last day of modelling time period as \code{Date}- object}

\item{path_std}{path to standard locations directory}

\item{path_climdb}{path to climate-db directory}

\item{dbout_name}{full path and file name for output-climate-db directory}
}
\value{
creates as SQLite-Database with one climate-df for each ID ind \code{df.ids$ID_custom})
}
\description{
This function creates a Project Climate Database in the format required by \code{\link[LWFBrook90R]{run_LWFB90}}. \code{\link[LWFBrook90R]{run_multisite_LWFB90}} can read climate data from this newly created database providing the advantage of lesser RAM consumption.
}
\examples{
# write project climate database from "2010-01-01" to "2011-12-31"
fnc_write_climdb(df.ids = test.ids.bds,
                 path_bdout = "H:/FVA-Projekte/P01540_WHHKW/Daten/Urdaten/test/")

# set clim_args
clim_args <-
  lapply(test.ids.bds$ID_custom,
         function(x) list(IDs = as.character(x),
                          path_climdb = "H:/FVA-Projekte/P01540_WHHKW/Daten/Urdaten/test/clim_data.sqlite"))
names(clim_args) <- as.character(test.ids.bds$ID_custom)

soil.test <- fnc_get_soil(df.ids = test.ids.bds,
                          soil_option = "BZE",
                          testgebiet = "BDS",
                          PTF_to_use = "HYPRES",
                          rootsmethod = "hartmann",
                          humus_roots = F)

res <- run_multisite_LWFB90(options_b90 = set_optionsLWFB90(startdate = as.Date("2010-01-01"),
                                                            enddate = as.Date("2011-12-31"),
                                                            root_method = "soilvar"),
                            param_b90 = fnc_get_params(tree_species = "spruce",
                                                       df.ids = test.ids.bds),
                            soil = soil.test,
                            climate = fnc_read_climdb,
                            climate_args = clim_args)
}
