% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnc_write_to_sql.R
\name{fnc_write_to_sql}
\alias{fnc_write_to_sql}
\title{Function to write temporary Outputfiles to SQLite-DB}
\usage{
fnc_write_to_sql(dir_tmp, db_name, del_tmp = T)
}
\arguments{
\item{dir_tmp}{path to temporary files}

\item{db_name}{name and file path of the SQL-database}

\item{del_tmp}{optional, shall tmp-files be deleted. Default is T}
}
\value{
writes the output to a database
}
\description{
Parallel writing is not supported for SQLite. To end with a SQLite-DB \code{\link{fnc_write}} and \code{\link{fnc_write_agg}} now create temporary .RData files that this function reads, writes to an SQLite-Database, and then deletes.
}
