% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnc_guessFK.R
\name{guess_fk}
\alias{guess_fk}
\title{FK estimate function}
\usage{
guess_fk(vals, days_passed = 3, perzentil = 0.95, with_image = NA)
}
\arguments{
\item{vals}{time series of soil moisture data given as vector}

\item{days_passed}{there are debates, whether field capacity is reached after 3,4, or 5 days. Can be set here. Default is 3 days.}

\item{perzentil}{exclude maxima outside of percentile}

\item{with_image}{whether guessing process shall be visualised. If so, provide a file name with full path.}
}
\value{
Value of volumetric water content (theta) for given pressure head (psi)
}
\description{
This simple function guesses field capacity (FK) based on a series of soil moisture values. The function looks for local maxima, takes the values \code{days_passed} after those local maxima and calculates a median of these values.
}
