% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MvG.R
\name{MvG.swc}
\alias{MvG.swc}
\title{Little MvG helper function}
\usage{
MvG.swc(psi, alpha, n, ThS, ThR, m = 1 - 1/n)
}
\arguments{
\item{SWATi}{swati from swatday.asc}

\item{soil}{soil.df}
}
\value{
Returns the desired output to the database directly.
}
\description{
...
}
