% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrecDailyToMonthly.R
\name{Prec.DailyToMonthly}
\alias{Prec.DailyToMonthly}
\title{Internal Helper Prec.DailyToMonthly}
\usage{
Prec.DailyToMonthly(dat)
}
\arguments{
\item{...}{}
}
\value{
Returns the desired output
}
\description{
...
}
