% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrecDailyToVegper.R
\name{Prec.DailyToVegper}
\alias{Prec.DailyToVegper}
\title{Internal Helper Prec.DailyToVegper}
\usage{
Prec.DailyToVegper(dat, vp.year, vp.start, vp.end, bypar)
}
\arguments{
\item{...}{}
}
\value{
Returns the desired output
}
\description{
...
}
