% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrecDailyToYearly.R
\name{Prec.DailyToYearly}
\alias{Prec.DailyToYearly}
\title{Internal Helper Prec.DailyToYearly
...}
\usage{
Prec.DailyToYearly(dat)
}
\arguments{
\item{...}{}
}
\value{
Returns the desired output
}
\description{
Internal Helper Prec.DailyToYearly
...
}
