% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SWATProfileDailyToMonthly.R
\name{SWATProfile.DailyToMonthly}
\alias{SWATProfile.DailyToMonthly}
\title{Internal Helper SWATProfile.DailyToMonthly}
\usage{
SWATProfile.DailyToMonthly(dat)
}
\arguments{
\item{...}{}
}
\value{
Returns the desired output
}
\description{
...
}
