% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SWATProfileDailyToYearly.R
\name{SWATProfile.DailyToYearly}
\alias{SWATProfile.DailyToYearly}
\title{Internal Helper SWATProfile.DailyToYearly
...}
\usage{
SWATProfile.DailyToYearly(dat)
}
\arguments{
\item{...}{}
}
\value{
Returns the desired output
}
\description{
Internal Helper SWATProfile.DailyToYearly
...
}
