% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnc_get_clim.R
\name{fnc_get_clim}
\alias{fnc_get_clim}
\title{Creation of Climate Data}
\usage{
fnc_get_clim(
  df.ids,
  mindate = as.Date("2010-01-01"),
  maxdate = as.Date("2011-12-31"),
  path_std = "R:/klima/whh/brook90_input/locations",
  path_climdb = "R:/klima/whh/brook90_input/db/",
  as = "list",
  db_path = NA
)
}
\arguments{
\item{df.ids}{a data frame containing the following columns:
\itemize{
\item \code{ID_custom} - a unique ID-column for assignment that all intermediate products as well as the output will be assigned to.
\item \code{easting} and \code{northing} - coordinates in UTM EPSG:32632
}}

\item{mindate}{first day of modelling time period as \code{Date}- object}

\item{maxdate}{last day of modelling time period as \code{Date}- object}

\item{path_std}{path to standard locations directory}

\item{path_climdb}{path to climate-db directory}

\item{as}{one of \code{list} or  \code{db}. Whether climate data should be written to a db, or kept in the R-sessions memory. If written to db, one has to handle db-querys in order to examine climate data. Yet, should be set to \code{db} the more points and years you model, as memory space fills quickly through climate data.}
}
\value{
Returns a named list (names from \code{df.ids$ID_custom}) of climate data
}
\description{
This function creates Climate Data in the format required by \code{\link[LWFBrook90R]{runLWFB90}} and \code{\link[LWFBrook90R]{msiterunLWFB90}}.
}
\examples{
fnc_get_clim(df.ids = test.ids.bds, mindate = as.Date("2010-01-01"), maxdate = as.Date("2011-12-31"))

}
