% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnc_get_clim.R
\name{fnc_get_clim}
\alias{fnc_get_clim}
\title{PTF application and humus layer creation}
\usage{
fnc_get_clim(
  df.ids,
  mindate = as.Date("2010-01-01"),
  maxdate = as.Date("2011-12-28"),
  path_std = "R:/klima/whh/brook90_input/locations",
  path_climdb = "R:/klima/whh/brook90_input/db/"
)
}
\arguments{
\item{df.ids}{a data frame containing the following columns:
\itemize{
\item \code{ID_custom} - a unique ID-column for assignment that all intermediate products as well as the output will be assigned to.
\item \code{easting} and \code{northing} - coordinates in UTM EPSG:32632
}}

\item{mindate}{first day of modelling time period as \code{Date}- object}

\item{maxdate}{last day of modelling time period as \code{Date}- object}

\item{path_std}{path to standard locations directory}

\item{path_climdb}{path to climate-db directory}
}
\value{
Returns a named list (names from \code{df.ids$ID}) of climate data
}
\description{
This function takes the data frame of soil physics data and creates the hydraulic parameters. It further creates humus-layers using the MvG-parameters from Hammel&Kennel (2001)
}
\examples{

fnc_get_clim(df.ids = test.ids.bds, mindate = as.Date("2010-01-01"), maxdate = as.Date("2011-12-31"))
}
