% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnc_PTF.R
\name{fnc_PTF}
\alias{fnc_PTF}
\title{PTF application and humus layer creation}
\usage{
fnc_PTF(df, PTF_used)
}
\arguments{
\item{df}{data frame containing information on soil physics in the following columns:
\itemize{
\item sand, silt, clay - soil texture in mass \%
\item bd - bulk density in g cm-1
\item oc.pct - organic carbon in mass \%
\item humus - thickness of the humus-layer (repeated \code{nrow} - times. a bit clumsy, but building on existing code it was the most convenient way)
}}

\item{PTF_used}{PTF-options from the \code{LWFBrook90R} - package. Choices are \code{"HYPRES"}, \code{"PTFPUH2"}, or \code{"WESSOLEK"}.}
}
\value{
Returns a longer data.table that already includes an earlier version of ls.soils. Further processed in \code{\link{fnc_get_soil}}.
}
\description{
This function takes the data frame of soil physics data and creates the hydraulic parameters. It further creates humus-layers using the MvG-parameters from Hammel&Kennel (2001)
}
\references{
Hammel, K., & Kennel, M. (2001). Charakterisierung und Analyse der Wasserverfügbarkeit und des Wasserhaushalts von Waldstandorten in Bayern mit dem Simulationsmodell BROOK90. Frank.
}
