% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnc_PTF.R
\name{fnc_PTF}
\alias{fnc_PTF}
\title{PTF application and humus layer creation}
\usage{
fnc_PTF(df, PTF_used, limit_humus)
}
\arguments{
\item{df}{data frame containing information on soil physics in the following columns:
\itemize{
\item sand, silt, clay - soil texture in mass \%
\item bd - bulk density in g cm-1
\item oc.pct - organic carbon in mass \%
\item humus - thickness of the humus-layer (repeated \code{nrow} - times. a bit clumsy, but building on existing code it was the most convenient way) \cr LWFB90 has problems if the first row is too thin and creates surface flow if it becomes saturated, which can happen if the humus layer is too thin. So, if humus is <= 2cm, it will be removed, 3 cm will be turned into 4 cm. Anything above is fine.
}}

\item{PTF_used}{PTF-options from the \code{LWFBrook90R} - package. Choices are \code{"HYPRES"}, \code{"PTFPUH2"}, or \code{"WESSOLEK"}.}

\item{limit_humus}{if \code{ilayer = 0} and \code{infexp = 0}, which is the default setting in \code{\link[LWFBrook90R]{set_paramLWFB90}}, water can only infiltrate in the first layer. If thin humus layers (e.g. 2 cm) are put on top of the first mineral soil layer, there's a risk of creating unrealisic amounts of surface runoff, when the water storage capacity of the thin humus layer can't take the incoming amount of daily rain. Hence, this parameter deletes humus lazers smaller or equal 2 cm, and increases humus layers of 3 cm to 4 cm. Default is \code{T}. If  \code{ilayer} and \code{infexp} are greater than 0, this issue shouldn't occur and this parameter should be set to \code{F}}
}
\value{
Returns a longer data.table that already includes an earlier version of ls.soils. Further processed in \code{\link{fnc_get_soil}}.
}
\description{
This function takes the data frame of soil physics data and creates the hydraulic parameters. It further creates humus-layers using the MvG-parameters from Hammel&Kennel (2001)
}
\references{
Hammel, K., & Kennel, M. (2001). Charakterisierung und Analyse der Wasserverfügbarkeit und des Wasserhaushalts von Waldstandorten in Bayern mit dem Simulationsmodell BROOK90. Frank.
}
