% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnc_read_climdb.R
\name{fnc_read_climdb}
\alias{fnc_read_climdb}
\title{Reading from Climate Database}
\usage{
fnc_read_climdb(IDs, path_climdb)
}
\arguments{
\item{path_climdb}{path to project climate data base}

\item{ID_custom}{ID-name of point in project climate data base}
}
\value{
returns climate data in correct format
}
\description{
This function reads in the Project Climate Database written by \code{\link[LWFBrook90R]{fnc_write_climdb}} and processes the data correctly, so that it can be used via the function parameters \code{climate} and \code{climate_args} in \code{\link[LWFBrook90R]{run_multisite_LWFB90}}. Function arguments must be passed to \code{\link[LWFBrook90R]{run_multisite_LWFB90}} as list of lists. Upper level containing the ID_custom names of the points to be modelled, the lower level containing the respective ID_custom name as well as the path to the project climate data base
}
\examples{
inst/examples/fnc_climdb_ex.R

}
