% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnc_roots.R
\name{fnc_roots}
\alias{fnc_roots}
\title{Fine-Root distribution}
\usage{
fnc_roots(df, rootsmethod = "hartmann", humus_roots = T, ...)
}
\arguments{
\item{df}{a data frame containing soil properties including MvG-alpha (column \code{alpha}), MvG-npar (column \code{npar}), theta_s (column \code{ths}), theta_r (column \code{thr}), and organic carbon in pct (column \code{oc.pct}.}

\item{rootsmethod}{name of the method for fine roots. Possible options are \code{hartmann},\code{betamodel}, \code{table}, \code{constant}, and \code{linear}. Default is \code{betamodel}.}

\item{humus_roots}{decides whether humus layers get roots too. Default is \code{TRUE}. If \code{TRUE}, humus layer gets the same amount of roots as highest soil layer, else \code{0}.}

\item{...}{additional input arguments that are passed on to \code{MakeRelRootDens}, see \code{\link[LWFBrook90R]{MakeRelRootDens}}.}
}
\value{
Returns the same data frame, but adds a rootden-column that is recognised by LWFBrook90R.
}
\description{
This function adds fine roots to the soil dataframes.
}
\references{
Hartmann, P., Buberl, H., Puhlmann, H., Schäffer, J., Trefz-Malcher, G., Zirlewagen, D., von Wilpert, K. (2016): Waldböden Südwestdeutschlands - Ergebnisse der Bodenzustandserhebungen im Wald von 1989 – 1992 und 2006 – 2008. Verlag Kessel, Remagen-Oberwinter, 328 S
}
