% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnc_soil_bze.R
\name{fnc_soil_bze}
\alias{fnc_soil_bze}
\title{Soil-list creation from BZE data}
\usage{
fnc_soil_bze(df.utm, df.assign, meta.out, ...)
}
\arguments{
\item{df.assign}{a dataframe containing the corresponding ID_custom for the IDs in \code{df.gk}}

\item{meta.out}{a string containing a path passed down from \code{fnc_get_soil}. Saving location of metadata.}

\item{...}{whether buffer should be used in extracting points from BZE raster files if \code{NAs} occur, options are \code{buffering} as \code{TRUE} or \code{FALSE}, and \code{buff_width} in \code{m}}

\item{df.gk}{A spatialpointsdataframe with the desired points in UTM25832.}
}
\value{
Returns a list of soil data frames.
}
\description{
This function is a wrapper of several smaller functions and chunks of code, all retrieved from "U:\\Brook90_2018\\paul_schmidt_walter_2018\\Dokumentation\\2_Bodenparameter.nb". Combining all this code into one function, it takes a spatialpointsdataframe of coordinates in GK-3 and returns a list of soil data frames. Those are further processed in \code{\link{fnc_get_soil}} by adding soil hydraulic information, humus, and fine roots and can then be read by \code{\link[LWFBrook90]{msiterunLWFB90}}.
}
