% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnc_soil_bze.R
\name{fnc_soil_bze}
\alias{fnc_soil_bze}
\title{Soil-list creation from BZE data}
\usage{
fnc_soil_bze(
  df.ids,
  meta.out,
  reduce_to_forest = F,
  limit_bodtief = NA,
  incl_GEOLA,
  buffering = F,
  buff_width = NA
)
}
\arguments{
\item{df.ids}{a dataframe with \code{ID_customs}, \code{IDs}, \code{aspect}, \code{slope}, \code{easting} and \code{northing}, as created one function up by \code{\link{fnc_get_soil}} May contain further columns}

\item{meta.out}{a string containing a path passed down from \code{fnc_get_soil}. Saving location of metadata.}

\item{limit_bodtief}{max soil depth, default is \code{NA} and uses max soil depth as defined in \code{df.LEIT}. If not \code{NA} soil-dfs are created down to the depth specified here as depth in \code{m}, negative. Might be used to give room for different \code{maxrootdepth} - settings in \link{fnc_get_params}. In this case, soil depth may be reduced significantly.}

\item{...}{whether buffer should be used in extracting points from BZE raster files if \code{NAs} occur, options are \code{buffering} as \code{TRUE} or \code{FALSE}, and \code{buff_width} in \code{m}}
}
\value{
Returns a list of soil data frames.
}
\description{
This function is a wrapper of several smaller functions and chunks of code, all retrieved from "U:\\Brook90_2018\\paul_schmidt_walter_2018\\Dokumentation\\2_Bodenparameter.nb". Combining all this code into one function, it takes a spatialpointsdataframe of coordinates in GK-3 and returns a list of soil data frames. Those are further processed in \code{\link{fnc_get_soil}} by adding soil hydraulic information, humus, and fine roots and can then be read by \code{\link[LWFBrook90]{run_multisite_LWFB90}}.
}
