% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnc_reduce.R
\name{fnc_write}
\alias{fnc_write}
\title{Function to reduce data output from automated LWFB90-Runs}
\usage{
fnc_write(x, dailycols, layercols, db_name)
}
\arguments{
\item{x}{one of the intermediate producs of \code{\link[LWFBrook90R]{run_LWFB90}} or  \code{\link[LWFBrook90R]{run_multisite_LWFB90}}, which is further processed internally. Can't be adjusted.}

\item{dailycols}{same as layercols but for daily output. For possible options see \code{\link[LWFBrook90R]{run_LWFB90}}}

\item{layercols}{a sting containing the desired output products. Full list of possible output columns can be find on the help page of \code{\link[LWFBrook90R]{run_LWFB90}} under \code{Layer outputs}}

\item{db_name}{name and file path of the SQL-database}
}
\value{
Returns the desired output to the database directly.
}
\description{
LWFBrook90 creates a lot of output files. In order to keep data storage to a minimum, both \code{\link[LWFBrook90R]{run_LWFB90}} and \code{\link[LWFBrook90R]{run_multisite_LWFB90}} provide an \code{output_fun} - argument that can be used to reduce the output and directly write it to a database. This is what this function is made for.
}
