% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnc_write_agg.R
\name{fnc_write_agg}
\alias{fnc_write_agg}
\title{Function to aggregate and write data from automated LWFB90-Runs}
\usage{
fnc_write_agg(x, aggr_tp, col_select_vp = NA, col_select_mon = NA, dir_name)
}
\arguments{
\item{x}{one of the intermediate producs of \code{\link[LWFBrook90R]{run_LWFB90}} or  \code{\link[LWFBrook90R]{run_multisite_LWFB90}}, which is further processed internally. Can't be adjusted.}

\item{aggr_tp}{a string containing the desired aggregation time period. Can be \code{monthly},  \code{vegper}, or  \code{monthly_vegper}. The latter does both.}

\item{col_select_vp}{a string containing the desired columns from the vegperiod-aggregation (see details)}

\item{col_select_mon}{a string containing the desired columns from the monthly-aggregation (see details)}

\item{dir_name}{directory for tmp files}
}
\value{
Returns the desired output to .Rdata files
}
\description{
LWFBrook90 creates a lot of output files. In order to keep data storage to a minimum, both \code{\link[LWFBrook90R]{run_LWFB90}} and \code{\link[LWFBrook90R]{run_multisite_LWFB90}} provide an \code{output_fun} - argument that can be used to reduce the output and directly write it to a database. This is what this function is made for. \cr In comparison to \code{\link{fnc_write}}, which only reduces the columns returned by \code{\link[LWFBrook90R]{run_LWFB90}} (see help page), this function enables aggregation over vegperiod and monthly, plus a more detailed selection of drought indices. See detail section.\cr\cr IMPORTANT: FOR RUNNING THIS AGGREGATE FUNCTION, \code{output} in \code{run_multiside_LWFB90} MUST BE SET TO A \code{df.output} AS SET BY THE CODE IN THE EXAMPLE SECTION \cr The function writes .RData files with the desired output for each point. \code{\link{fnc_write_to_sql}} can be used to convert these files into a SQLite-DB. This "step-in-between" is necessary because SQLite does not support parallel writing.
}
\section{Vegperiod and monthly outputs}{

For a complete list of possible output types plus description, see \code{"U:/db_brook90_output/whh_db_documentation"}
}

\examples{
df.output <- set_outputLWFB90()
df.output[,] <- 0L
df.output[c("Evap", "Swat"), c("Day")] <- 1
df.output[c("Flow"), c("Mon")] <- 1

}
