% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnc_write_climdb.R
\name{fnc_write_climdb}
\alias{fnc_write_climdb}
\title{Creation of Climate Database}
\usage{
fnc_write_climdb(
  df.ids,
  mindate = as.Date("2010-01-01"),
  maxdate = as.Date("2011-12-31"),
  path_std = "R:/klima/whh/brook90_input/locations",
  path_climdb = "R:/klima/whh/brook90_input/db/",
  path_bdout
)
}
\arguments{
\item{df.ids}{a data frame containing the following columns:
\itemize{
\item \code{ID_custom} - a unique ID-column for assignment that all intermediate products as well as the output will be assigned to.
\item \code{easting} and \code{northing} - coordinates in UTM EPSG:32632
}}

\item{mindate}{first day of modelling time period as \code{Date}- object}

\item{maxdate}{last day of modelling time period as \code{Date}- object}

\item{path_std}{path to standard locations directory}

\item{path_climdb}{path to climate-db directory}

\item{path_dbout}{path to output-climate-db directory, i.e. project directory}
}
\value{
creates as SQLite-Database with one climate-df for each ID ind \code{df.ids$ID_custom})
}
\description{
This function creates a Project Climate Database in the format required by \code{\link[LWFBrook90R]{run_LWFB90}}. \code{\link[LWFBrook90R]{run_multisite_LWFB90}} can read climate data from this newly created database providing the advantage of lesser RAM consumption.
}
\examples{
inst/examples/fnc_climdb_ex.R

}
